+++ package.json
@@ -2,9 +2,11 @@
   "name": "nexus-validation-plugin",
   "version": "0.0.1",
   "description": "Nexus plugin for argument validation.",
-  "main": "index.js",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
   "scripts": {
     "build": "tsc",
+    "prepare": "npm run build",
     "lint": "eslint --ext .ts,.tsx,.js,.jsx src",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
